+++ examples/ssr/src/App.tsx
@@ -1,4 +1,3 @@
-import * as React from "react";
 import { Routes, Route, Outlet, Link } from "react-router-dom";
 
 export default function App() {