+++ packages/react-router-dom-v5-compat/jest.config.js
@@ -21,5 +21,4 @@ module.exports = {
     "^react-router$": "<rootDir>/../react-router/index.ts",
     "^react-router-dom-v5-compat$": "<rootDir>/index.ts",
   },
-  reporters: ["default"],
 };