+++ packages/react-router-dom/jest.config.js
@@ -21,5 +21,4 @@ module.exports = {
     "^react-router$": "<rootDir>/../react-router/index.ts",
     "^react-router-dom$": "<rootDir>/index.tsx",
   },
-  reporters: ["default"],
 };