+++ packages/router/__tests__/browser-test.ts
@@ -10,6 +10,7 @@ import Listen from "./TestSequences/Listen";
 import PushNewLocation from "./TestSequences/PushNewLocation";
 import PushSamePath from "./TestSequences/PushSamePath";
 import PushState from "./TestSequences/PushState";
+import PushStateInvalid from "./TestSequences/PushStateInvalid";
 import PushMissingPathname from "./TestSequences/PushMissingPathname";
 import PushRelativePathname from "./TestSequences/PushRelativePathname";
 import ReplaceNewLocation from "./TestSequences/ReplaceNewLocation";
@@ -22,10 +23,11 @@ import ListenPopOnly from "./TestSequences/ListenPopOnly";
 
 describe("a browser history", () => {
   let history: BrowserHistory;
+  let dom: JSDOM;
 
   beforeEach(() => {
     // Need to use our own custom DOM in order to get a working history
-    const dom = new JSDOM(`<!DOCTYPE html><p>History Example</p>`, {
+    dom = new JSDOM(`<!DOCTYPE html><p>History Example</p>`, {
       url: "https://example.org/",
     });
     dom.window.history.replaceState(null, "", "/");
@@ -90,6 +92,10 @@ describe("a browser history", () => {
   describe("push state", () => {
     it("calls change listeners with the new location", () => {
       PushState(history);
+    });
+
+    it("re-throws when using non-serializable state", () => {
+      PushStateInvalid(history, dom.window);
     });
   });