+++ packages/react-router/CHANGELOG.md
@@ -1,5 +1,16 @@
 # `react-router`
 
+## 6.12.0
+
+### Minor Changes
+
+- Wrap internal router state updates with `React.startTransition` if it exists ([#10438](https://github.com/remix-run/react-router/pull/10438))
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/router@1.6.3`
+
 ## 6.11.2
 
 ### Patch Changes